WIP v0.3.0 add primitives/binary_tools
